+++ src/argv.c
@@ -4,36 +4,40 @@
 #include <string.h>
 
 struct _Argv {
-	char **_list;
-	int _max, _n;
+	char **list;
+	int max, n;
 };
 
-char *argv_strlist( Argv *argv ) {
+void argv_strlist( Argv *argv, char *dest, size_t size ) {
 
 	int i, len;
 	char *_slist;
 
 	if ( !argv )
-		return NULL;
+		return;
 
 	len = 0;
 
-	for ( i=0; argv->_list[ i ]; i++ ) {
-		len+=strlen( argv->_list[ i ] )+1;
+	for ( i=0; argv->list[ i ]; i++ ) {
+		len+=strlen( argv->list[ i ] )+1;
 	}
 
 	_slist = str_alloc( len );
 
 	if ( _slist ) {
+
 		_slist[ 0 ] = '\0';
-		for ( i=0; argv->_list[ i ]; i++ ) {
-			strcat( _slist, argv->_list[ i ] );
+
+		for ( i=0; argv->list[ i ]; i++ ) {
+			strcat( _slist, argv->list[ i ] );
 			strcat( _slist, " ");
 		}
 
 	}
 
-	return _slist;
+	strncpy( dest, _slist, size );
+	str_destroy( _slist );
+
 }
 
 void argv_clean( Argv *argv ) {
@@ -42,11 +46,11 @@ void argv_clean( Argv *argv ) {
   if ( !argv )
 		return;
 
-	if ( argv->_list ) {
+	if ( argv->list ) {
 
-    for ( i=0; i < argv->_n; i++ ) {
-			free( argv->_list[ i ] );
-			argv->_list[ i ] = NULL;
+    for ( i=0; i < argv->n; i++ ) {
+			free( argv->list[ i ] );
+			argv->list[ i ] = NULL;
 		}
 
   }
@@ -60,9 +64,9 @@ void argv_free( Argv *argv ) {
 
   argv_clean( argv );
 
-  if ( argv->_list ) {
-    free( argv->_list );
-    argv->_list = NULL;
+  if ( argv->list ) {
+    free( argv->list );
+    argv->list = NULL;
   }
 
 	free( argv );
@@ -74,18 +78,21 @@ Argv* argv_build( int q ) {
 	int i;
 	Argv *argv;
 
+	if ( q <= 0 )
+		return NULL;
+
 	argv = (Argv*) malloc( sizeof( Argv ) );
 
 	if ( argv ) {
-		argv->_list = (char**)malloc( q*sizeof( char* ) );
-		if ( argv->_list ) {
+		argv->list = (char**)malloc( q*sizeof( char* ) );
+		if ( argv->list ) {
 
 			for (i=0; i<q; i++) {
-				argv->_list[ i ] = NULL;
+				argv->list[ i ] = NULL;
 			}
 
-			argv->_n = 0;
-			argv->_max = q;
+			argv->n = 0;
+			argv->max = q;
 
 		} else {
 			argv_free( argv );
@@ -97,22 +104,32 @@ Argv* argv_build( int q ) {
 
 void argv_add( Argv *argv, const char *str ) {
 
-	int len;
 	if ( !argv || !str )
 		return;
 
-	if ( argv->_n+1 < argv->_max ) {
-		len = strlen( str );
-		argv->_list[ argv->_n ] = str_alloc( len );
-		strcpy( argv->_list[ argv->_n ], str );
-		argv->_n++;
+	if ( argv->n < argv->max ) {
+		argv->list[ argv->n ] = str_clone( str );
+		argv->n++;
 	}
 
 }
 
-char** argv_get_list( Argv *argv ) {
-	if ( !argv )
+char *argv_get_idx( Argv *argv, int idx ) {
+
+	if ( !argv || idx < 0 )
 		return NULL;
 
-	return argv->_list;
+	if ( idx >= argv->n )
+		return NULL;
+
+	return argv->list[ 0 ];
+
+}
+
+int argv_n( Argv *argv ) {
+
+	if ( !argv )
+		return -1;
+
+	return argv->n;
 }